MongoDB

+++ back/resetdb.mjs
@@ -4,8 +4,11 @@ const client = new MongoClient('mongodb://localhost:27017', { useUnifiedTopology
 try {
   await client.connect();
   const db = client.db('inno-comp');
-  let collection = db.collection('applicants');
-  collection.drop();
+  const collections = await db.collections();
+  let collection = collections.find(c => c.collectionName === 'applicants');
+  if (collection) {
+    collection.drop();
+  }
   collection = await db.createCollection('applicants');
   console.log(collection.dbName);
   console.log(collection.collectionName);
